requirements.txt
# Web framework and dependencies
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.2

# Deployment dependencies
psycopg2-binary==2.9.6
